Added EntityDrawer and unfolded components info is now shared between entities from same pool
