+++ Entitas.Unity.VisualDebugging/Assets/Examples/SystemsController.cs
@@ -8,8 +8,8 @@ public class SystemsController : MonoBehaviour {
     Pool _pool;
 
     void Start() {
-        _pool = new Pool(ComponentIds.TotalComponents, 0, new PoolMetaData("Systems Pool", ComponentIds.componentNames));
-        new PoolObserver(_pool, ComponentIds.componentTypes);
+        _pool = new Pool(ComponentIds.TotalComponents, 0, new PoolMetaData("Systems Pool", ComponentIds.componentNames, ComponentIds.componentTypes));
+        new PoolObserver(_pool);
         _systems = createNestedSystems();
         _systems.Initialize();
         _pool.CreateEntity().AddMyString("");
